Update expected behavior for WiFi scan async result
Update documentation in the CHRE API and PAL to modify the stated
expected behavior surrounding the async result of an explicit scan
request. This result must only come after the final result of the scan
is known, and not at the point where the scan is scheduled but may still
fail.

Change-Id: Ic7acc45038c6ba1dfe81e7ee5c586c2e78ecc088