Add nanoapp discovery mechanism to CHRE API
Introduce a new API, chreGetNanoappInfo(), which can be used to retrieve
the instance ID for a given nanoapp ID. Also add events which inform
nanoapps of the starting and stopping of other nanoapps.

Also, define the vendor IDs for Google and provide a helper macro to
mask off the vendor ID portion of nanoapp or CHRE platform IDs.

Change-Id: Ib04c5db64e0245ad23c560daeb307a10e963d323