Add initial preview of CHRE API v1.1
Introduce the current preview release of the next version of the CHRE
API, version 1.1. This new version builds upon v1.0 by adding support
for the following features/capabilities:

 - Ability to request GNSS-based position fixes, along with raw
   measurement data
 - Ability to register for monitoring the results of WiFi scans, and
   to explicitly request a new WiFi scan
 - Ability to query for cell identity information from the WWAN modem,
   including information about neighbor cells
 - Ability to lookup the instance ID of other nanoapps running in the
   system, enabling nanoapps to provide services for other nanoapps

This version of the API also extends and updates the top-level
documentation in chre.h, covering platform expectations surrounding
compatibility, threading, etc.

As of this CL, the API is in preview stage and is still subject to
change.

Test: compile test only
Change-Id: I2baf93130f1ef03b1a4c94d4465f64c1010dd4ea