Add deferCallback() for use by the CHRE system
Leverage the event subsystem to implement a method of invoking a
callback from within the context of an EventLoop at some point in the
future. This can be used to transfer control from a potentially
arbitrary thread in a platform-specific callback, into a known
EventLoop context, as well as to defer processing of some event.

Also, modify EventLoop::postEvent to indicate whether the event was
successfully posted to the queue.

Change-Id: I79e5b8d5aff82fd2fb304b2a1055aebc3047b50d