CHRE API: Add hostEndpoint concept, remove obsolete note
Introduces chreSendMessageToHostEndpoint(), which takes an additional
parameter to specify the entity on the host that should receive the
message, and changes the type of the messageSize parameter from
uint32_t to size_t to preserve symmetry with chreMessageFreeFunction.
Also updates the struct given when a message is received to indicate
the sending endpoint.

Removes a note in chreSendEvent mentioning that other nanoapps' instance
IDs are not easily discoverable, as this is now possible via
chreGetNanoappInfo().

Change-Id: I30b49052c900a51840656ebf243e334f9bfd3d2b