Add SocketClient, move shareable code to common
Add a SocketClient as a helper to simplify interacting with the socket
connection, including automatically reconnecting if the remote end dies.
Another layer is expected to be added on top of this to handle
encoding/decoding messages.

Move the SocketServer and chre_test_client code from the MSM-specific
folder to common, as these can be used in any platform where there's a
userspace daemon that acts as the entry point to CHRE communications.

Test: compile, run chre and chre_test_client
Change-Id: I69c06eb823670beb0a8d018b8fe80495ced842a5