Add HostProtocol for common encode/decode
Add a new helper class called HostProtocol which takes care of encoding
and decoding FlatBuffers messages. HostProtocol can be used both in the
CHRE and in host-side code.

Update the host messages IDL to mark two fields as required, which
removes the requirement for code to check that these fields are not null
when handling a decoded message. If the field is ommitted, the message
will be rejected at the verification step.

Rework HostLink and chre_test_client to leverage HostProtocol. Reorder
hostEndpoint and messageType parameters in HostCommsManager to ensure
consistency.

Test: load CHRE, run chre_test_client with MessageWorld
Change-Id: I6221d78f881bb7fe29f9fbe8a842ca2b07e5674d