Rework Nanoapp model
Unify Nanoapp into a single monolithic object that directly exposes the
common interfaces implemented in core and platform. As a result of this,
PlatformNanoapp will no longer exist as a separate entity from Nanoapp.
In preparation for dynamic app support, refactor static nanoapp loading
to construct a UniquePtr<Nanoapp> and transfer ownership to the
EventLoop when starting a nanoapp.

This commit includes the first look at the nanoapp support library which
will be used for external nanoapps implemented as dynamic shared
objects.

Test: run chre_test_client with MessageWorld, and run the Linux simulator
Change-Id: Ic17f2c48c3156d1e5bb255b1c6a188908123d5df