Move unique pointer construction to free function
Including a forwarding constructor in UniquePtr prevents creation of a
move constructor, which is necessary to resolve an issue in the
containers. Instead, use a new free function called MakeUnique to
perform allocation + in-place construction + assignment to a UniquePtr.
Also, add new UniquePtr constructors that enable (1) default
initialization to null, (2) assume ownership of a pointer to dynamically
allocated memory, and (3) transfer ownership from another UniquePtr.

Test: run simulator and tests
Change-Id: Id8f91d7452ccb19a4bc7acf3ad87e8e2db3bac5f