Update HAL to pass VTS, increase safety
Make client callbacks thread-safe by adding a lock that protects against
concurrent modification of the callbacks structure and dereferences to
invoke a callback. Fix a bug in queryApps that resulted in it always
returning failure even in the successful case. Change the return value
of unload/enable/disable to TRANSACTION_FAILED to match expectations
from VTS.

Test: run VtsHalContexthubV1_0TargetTest binary from VTS directly
Change-Id: I6044d805cd20cc39ef7e02da9501f37ad05ad960