Leave reservedMessageType as a deprecated alias
Preserve the reservedMessageType name used in CHRE API v1.0 as an alias
to the field that has now been renamed to messageType. This allows code
that references it to compile and function normally against both v1.0
and v1.1 API definitions.

This alias is deprecated and will be removed in a future update.

Test: build CHRE
Change-Id: I2e18ab9359bb2b7c55c0b3090c88d0d852493d1d