Updates to Optional for safety and consistency
After the destructor is invoked via Optional::reset(), the underlying
object is invalid and cannot be safely used for assignment. Avoid this
by performing move/copy construction in the Optional assignment
implementations if the underlying storage is uninitialized.

Also, bring chre::Optional in-line with std::optional by removing the
requirement for the contained object to have a default constructor
(instead leaving the object storage uninitialized after default
construction of Optional), and making the semantics of move assignment
the same (only the underlying object is moved, not the "ownership" of
the object, i.e. the moved-from Optional can still have has_value() set
to true).

Test: ./run_tests.sh
Change-Id: Ie3627838aee10acf89d3cf609aebf8fc6694ebea