Expose the estimated host time offset in the CHRE API
Rather than only providing a function that gives the current estimate of
the host's time, add a new API called chreGetEstimatedHostTimeOffset()
which supplies the current estimate of the offset. Convert
chreGetEstimatedHostTime() into a convenience function that applies this
offset to the current CHRE time.

Also, relax the host time accuracy requirement to +/- 10ms.

Test: compile simulator
Change-Id: I73886355bd1b2efc68dbbedd25b1d2979b0854bb