Increase max message to host size, minor cleanup
On SLPI, the host allocates a 4KB buffer to use when receiving messages
from CHRE. Increase the maximum message size we accept from nanoapps
from 2048 to 4000 bytes to allow nanoapps to use a larger portion of
this buffer. The remaining 96 bytes are reserved for current and future
overhead.

Also, avoid reallocation of the buffer used to construct the FlatBuffer
for nanoapp messages to the host by increasing the fixed buffer
reservation.

Drop log message on path from host to nanoapp in the daemon from debug
to verbose level, which matches the similar verbose log in the reverse
path.

Bug: 37640330
Test: run MessageWorld on-device + chre_test_client
Change-Id: I9dfccf7b00d3d09b3be7f529271432b10c2da044