Add debug dump support
Integrate CHRE with the new ASH debug dump implementation to enable
dumping debug info to a place where it will be included in a bugreport.
This is accomplished by implementing the debugDump function in the HIDL
interface in the context hub HAL, and sending a message to CHRE to
retrieve debugging information as a series of strings. CHRE currently
does not put any useful information in this log, only a test string.

Bug: 37640330
Test: load on device with updated SLPI firmware, run
  lshal debug android.hardware.contexthub@1.0::IContexthub/default
  also, collect a bugreport and confirm that the info shows up in
  lshal-debug.txt
Change-Id: Ia841d623b1e581d4435f907700082567edb53769