Define CHRE_FILENAME in build, use for fatal error
Supply the CHRE_FILENAME macro, which is supplied by the build system to
contain the filename without path information. Use this to create a
string at compile-time to identify the site where ERR_FATAL is called on
SLPI, so the crash signature is unique.

This produces a string in dmesg that looks like this:
"err_qdi.c:435:EF:M8998AAAAAAAAQ1:1:CHRE fatal@timer_world.cc:52."

Bug: 63171037
Test: Instrument TimerWorld to trigger a fatal error, check dmesg
Change-Id: I998225b743c34c39b4e0e217213c968b39427596