[IMU_CAL] Keep temp in passive mode, fix race
As a workaround for an issue with passive sensors in CHRE, revert the
change where IMU cal would disable the temperature sensor when gyro is
disabled. This exposed a sequence of events where CHRE would disable its
passive accel registration even when it's active for other clients.
Keeping the temperature sensor in passive mode (instead of toggling it
off/on) avoids this problem, and adjusting the latency to a long window
while in stand-by mode retains nearly all of the power benefits of the
previous implementation.

Avoid a potential race condition between receiving an event with
updated gyro sampling status and reading the current state by using the
state from the sampling status change event rather than querying.

Bug: 63844556
Bug: 63908396
Test: confirm that IMU cal receives samples for all expected sensors
when sensortest is used to register accel+gyro at the same time
Change-Id: Iffdea60a94d52e2fcf0c2bfae69c6699cb015e73