Mark chre_api headers as a vendor module
While this API definition is shared to the ecosystem, the choice of API
version and implementation thereof is device-specific. Therefore, this
module should be tagged as a vendor module, which also allows it to be
used by other vendor modules.

Bug: 32842314
Test: mma from system/chre path
Change-Id: I0a27878a47d3b235736e238a77441d4256e63fdd
Merged-In: I0a27878a47d3b235736e238a77441d4256e63fdd