Avoid deadlock with SMR
SMR holds the client mutex when invoking response callbacks, and also
acquires it when issuing requests. CHRE previously acquired a mutex
prior to issuing a request, and also when handling a response. This
interaction could lead to deadlock if a request times out.

Address this by (1) increasing the default SMR request timeout from 1 to
2 seconds, and (2) refactoring chre::SmrHelper to not hold a mutex when
issuing a request. As part of (2), also update SmrHelper to free memory
associated with a timed out request if the callback is received after
the timeout.

Bug: 68862548
Test: force timeout by setting timer very low, monitor logs to ensure
expected behavior arises, also use SensorWorld to check basic sensor
sanity

Change-Id: Ib7c6743dc2c0b152fe4bc782c2a73eedfdfa2f24