CHRE API v1.2: Add WiFi radio chain preference
Introduce a new parameter in the on-demand scan request that allows
nanoapps to specify their desired trade-off with regards to use of
multiple radio chains when performing a scan. This preference can be
used by the WiFi firmware to decide whether to perform a Dual Band
Simultaneous (DBS) scan, for example, which can save power at the
expense of RSSI accuracy.

Test: compile only
Change-Id: I1de407c6930d754cd1f114b3a3f44010d09a2c55