Fix backwards compatibility lookup
References to weak symbols defined within the same .so always call
through to the weak definition, rather than using the undefined symbol
resolution process. So update the weak symbol stubs for newly defined
functions to dynamically resolve the address of the platform's function
definition, and call through to it (if present).

Bug: 73540991
Test: build a nanoapp that references new v1.2 APIs and confirm it calls
through to CHRE when running on a v1.2 platform, and exhibits default
behavior when that function is not provided by the platform
Change-Id: I7306ca5cd1c57219cfcad171363a6a41cf3f8890