Fix usages of [un]registerForBroadcastEvent
These functions actually return false if no change was made to the
nanoapp's event registrations, and not on failure. Fix usages of this
function that made this assumption, especially those that treat a false
return value as an error (which is already done in the case of failing
to register for an event).

Test: call chreConfigureHostSleepStateEvents(false) from nanoappStart
Change-Id: I6faf51eecfabdc9b8356db2a85f707c33c7f90c6