Define host messages for audio request
Add new empty message types to be used to notify the host when the CHRE
has an active request for audio data, along with encode helpers and a
hook for handling them in the daemon. These messages will be used to
interact with other host-side components to ensure the necessary
hardware is enabled on platforms that require such coordination.

Bug: 67001170
Test: compile only
Change-Id: I519dd6c5d81d13f59b7b414230716f50f52fbfda