Extract calibration management into SeeCalHelper
Move calibration-related code out of SeeHelper and into a new class
called SeeCalHelper, and allow multiple instances of SeeHelper to share
the same SeeCalHelper object. This allows us to handle calibration
updates within the micro-image instance of SeeHelper, and used the
cached data to apply calibration to raw samples received in both
instances, avoiding the need to transition to big image each time
calibration data is updated.

Bug: 79257265
Test: with additional logging enabled, confirm that we stay in
micro-image when calibration updates are generated for accel at 1 Hz,
and SensorWorld running in big image receives samples with the offsets
applied
Change-Id: I123175063eac7388834ceadc775df6cd84191083