Undo island transitions during blocking SEE calls
If CHRE gets blocked waiting on a response from SEE, during which we
transition from big image to micro-image (island mode), force a
transition back to big image.

Although this won't happen if the transition occurs due to a request
made by a big image nanoapp (we hold an island block while it is
executing), other areas of the code may assume that transitions to
micro-image only happen when the CHRE thread is waiting on events.

Bug: 111066017
Test: sanity check nanoapp functionality
Change-Id: I419b0491a85f57a23879e8e12dcd0eb009814754