CHQTS: Moves passive request checks to end of test
The current test uses a self-event as a divider between events
that occur due to passive/active requests. However, there is no guarantee
that we won't receive events that come from passive requests after
the self-event. In some cases, this causes the nanoapp to receive
unexpected disable events, causing a test failure.

To avoid passive requests from interfering with the rest of the test,
this CL moves the check to the end of the test (and removes the
self-event logic). Currently the test only checks for the return value,
so there is no loss in test coverage with this change.

Bug: 111444100
Test: Compile and run CHQTS, verify pass
Change-Id: I1871dce18ad46b20cbda016895f1d8da2df5fbc4