Fix race between unload and message receipt
For DSO-based nanoapps, metadata is held within the shared object and
used to return data about the app, like its app ID. On SLPI, this data
can also be accessed from another thread, which opens up a race
condition between unloading the nanoapp and sending a message to a
nanoapp, which involves iterating through the list of nanoapps.

Close this race by holding the nanoapp list lock before calling
PlatformNanoapp:end(), and until after the Nanoapp object is destructed.

Also, avoid other potential issues by setting mAppInfo to null prior to
calling dlclose(), so the soon-to-be unmapped memory can't be
referenced.

Bug: 117732008
Test: use chre_test_client to validate unload + load cycle 2x
Change-Id: Ie2b4d4f665d9a113b6f154c0478572500998219f