Mark GNSS stop updates on inactive requests as success
It's possible for a nanoapp to call chreGnssLocationSessionStopAsync
multiple times before the async response is processed by CHRE. In that case,
CHRE may crash since it will attempt to update the nanoapp request,
which is no longer active due to processing the former request. Per API
requirements, if no GNSS session was active at the time of the request,
CHRE should report as if it was successfully ended. To reflect this
behavior, mark stop updates for inactive nanoapp requests as a
success.

Bug: 123587718
Test: Test GNSS world and verify chreGnssLocationSessionStopAsync
      can be called multiple times.

Change-Id: I32e4e53bf9887584f866a4597a31b58d20e56b56