Allow stopping inactive WiFi monitor requests
It's possible for a nanoapp to call chreWifiConfigureScanMonitorAsync
multiple times before the async response is processed by CHRE.
In that case, CHRE may crash since it will attempt to update
the nanoappp request which is no longer active due to processing
the former request.

Also updates the API to reflect that this behavior is OK and
expected.

Bug: 130247334
Test: Load CHRE and a nanoapp that exhibits the above behavior
      and verify crashes don't occur.

Change-Id: Iae6b461abf8a654e9f18b7758bf92214ff3ca4a0