Add received time to Event.h
Adding the received time to events can help indicate if CHRE has been
stuck processing some event for a long period of time or if a stream
of events came in too quickly to process.

Only a uint16_t is used because having more than 65 seconds worth of
data would indicate CHRE is stuck so there isn't a need to store more
fine grained than that.

Size before: 101966
Size after: 102070

Test: Add log after event creation and ensure received
   time is recorded
Bug: 133164972

Change-Id: I0701132ac680b93647f81d574ba8e7090d693ff1