Init sensors after the CHRE thread has started
Today, CHRE blocks to initialize sensors as soon as it's initialized.
This means it's unable to serve the getHubInfo request that comes in
during boot until after all sensors come online. Sometimes sensors take
a while to come up (especially during the boot after a flash / update)
so it'd be good if the amount of time CHRE waits for sensors to come up
could be increased. This CL moves sensor initialization to lateInit()
so CHRE can serve the getHubInfo request and sensor init isn't blocking
boot.

Bug: 131330968
Test: ./load_chre.sh and reboot. Verified that all sensors come up
    after the CHRE thread starts. Tested on both SEE / SMGR devices.

Change-Id: I4db4c5f303ed8f9221e95f3358dd46f88a899f43