Increase SEE init timeout to 60 seconds
Now that sensor initialization doesn't block device boot, increase the
timeout we allow SEE to initialize to a fairly large number to avoid
cases where the device is slow after an upgrade or flash (for thermal
throttling reasons or otherwise). This allows distinguishing between a
system being slow or the sensors code being deadlocked.

Bug: 131330968
Test: compile
Change-Id: I3f7a8d3c79e60c910d40d5c7b2b9ef9350116918