Add retry logic when starting CHRE
Since the SLPI may not be ready yet when the CHRE daemon attempts to
start the CHRE thread, retry the first FastRPC method a few times before exiting
to avoid increased startup delay. Moves the initial time sync message call to
after starting the CHRE thread.

Bug: 135772692
Test: Push to device, verify that system startup time is not delayed.
Verify via logs that the initial time sync message is sent, and add
a debugging global ctor/dtor in the CHRE SLPI library to ensure that
the dynamic module is loaded only once.
Change-Id: Id873c29fd268b2ac501873cef5ea3e1bb3457a31