Fix power test nanoapp TCM build
While CHRE_SLPI_UIMG_ENABLED is defined in variants built for TCM, the
Makefile defining the flag isn't included until after it's checked for
in the power test Makefile.

Since it's not possible to use this flag, define two separate
directories containing a Makefile for the TCM and non-TCM versions of
the nanoapp and put the nanoapp code into a common directory that is
referenced by both Makefiles.

Bug: 139427998
Test: Build both TCM / non-TCM version of the nanoapp and verify they
    load with different nanoapp IDs

Change-Id: I033c0210e9f9b81d41d9dd2cf84671e1ca8a845f