Move implementation into ChreCrossValidatorManager class
Move all nanoapp functionality into the manager class and keep it mostly
the same. Only difference is that instead of having several optional
values that determine state of the validation move them all into one
optional CrossValidatorState struct. If this state does not have a value
set that will indicate error of setup.

Bug: 146052784
Test: Run accelerometer sensor data CHQTS test with validator
Change-Id: Ia2339bf8e3c230d08bda7c979ccb5f4df35a938e