Add CHRE->CHPP structure conversion
Recursively generate conversion functions from CHRE to CHPP sturctures,
for all structs, starting from the root struct(s). This is initially
done via generated field assignments for primitive, and function calls
for nested structures, to ensure conversion to the packed layout is
correct. In the future, this can be optimized to generate code that
checks at compile time whether the binary layouts are equivalent (sizeof
matches, offsetof for each field matches), and if so, just memcpy(),
ideally the minimal number of calls to memcpy().

Does not include code to allocate the memory for the target CHPP
structure yet, and doesn't handle unions yet.

Also includes updates to annotations to match recent CHRE API updates to
make the union in chreWwanCellInfo non-anonymous, and a preview of what
is planned to be used to properly map from the enum discriminator to the
appropriate conversion function.

Test: output looks reasonable, more thorough testing will come later
Change-Id: I75ca17cd56bc485ec74fece41b872b0bf4cd74a8