missing break statements in handleEvent switch-case
This was causing async result events to be passed to handl scan result
method and all event were triggering unrecognized event log. The handle
scan result call that was passed the async result struct was
interpreting the resultTotal field to be = to 0 before. So now that that
is removed the logic handling that error can also be removed.

Bug: 154271547
Test: Run wifi cross val chqts test
Change-Id: I27b0efed852328d2731e240a2cd034416f219e8e