Reduce verbosity of CHRE build system
Significantly reduces the verbosity of the build system by silencing the
output of the compilation and linking commands in the build template and
for nanopb and replaces them with short echo statements.

This helps make it easier to see what's being compiled and doesn't flood
the console.

In case verbosity is needed, a flag has been added to allow increased
verbosity.

Bug: 159742444
Test: Compile CHRE
Change-Id: Icd87ea3bd8eb468d943453e688f9fc7be3ac370e