Update postEventOrDie to return void
Since this function will trigger a fatal error on failure, checking the
return value is unnecessary and unexpected, so make it return void.
Also ensure that if an event is posted while CHRE is shutting down
(possible in a race condition), memory associated with the event is not
leaked.

Also do some minor refactoring in the TimerPool implementation to
improve clarity.

Fixes: 169393998
Test: ./run_sim.sh
Change-Id: I53081f2a0366d421d1f3cad7f2952329816f0639