Avoid copying uninitialized bytes in NestedDataPtr
Since we only initialize the number of bytes needed for the nested type,
we should do the same when populating a void* output value to avoid
confusing tools that check for use of uninitialized memory.

Also, throw in a static_assert to confirm that the nested type is
trivially copyable, as expected.

Bug: 175163228
Test: compile for coral, run msan test on google3
Change-Id: Ieaf94eeed12e930e974c9cd6b5f8034179791e1d